Moved choice_fast to its own file in common. Added license to noisers_test.py
